requirements.txt
fastapi[standard]
python-dotenv
pwdlib[argon2]
passlib[argon2]
pyjwt
bcrypt
fastexcel
SQLAlchemy
asyncpg
uvicorn
sqlacodegen
psycopg2
polars
fastexcel
python-calamine
rapidfuzz
pandas 
joblib
scipy
xgboost
asyncpg